General cleanup and fixing all kinds of log
