+++ src/concurrency/timestamp_ordering_transaction_manager.cpp
@@ -753,10 +753,7 @@ ResultType TimestampOrderingTransactionManager::CommitTransaction(
   //////////////////////////////////////////////////////////
 
   auto &manager = catalog::Manager::GetInstance();
-  auto &log_manager = logging::ReorderedPhyLogLogManager::GetInstance();
-
-  //log_manager.StartLogging();
-  //log_manager.RegisterWorker(current_txn->GetThreadId());
+  auto &log_manager = logging::WalLogManager::GetInstance();
 
   cid_t end_commit_id = current_txn->GetCommitId();
 
@@ -832,9 +829,9 @@ ResultType TimestampOrderingTransactionManager::CommitTransaction(
         // add old version into gc set.
         // may need to delete versions from secondary indexes.
         gc_set->operator[](tile_group_id)[tuple_slot] = GCVersionType::COMMIT_UPDATE;
-        log_manager.StartPersistTxn(101);
-        log_manager.LogUpdate(old_version, new_version);
-        log_manager.EndPersistTxn(101);
+        //log_manager.StartPersistTxn(end_commit_id);
+        log_manager.LogUpdate(old_version, new_version,5);
+        //log_manager.EndPersistTxn(end_commit_id);
       } else if (tuple_entry.second == RWType::DELETE) {
         ItemPointer new_version =
             tile_group_header->GetPrevItemPointer(tuple_slot);
@@ -864,9 +861,9 @@ ResultType TimestampOrderingTransactionManager::CommitTransaction(
         // recycle old version, delete from index
         // the gc should be responsible for recycling the newer empty version.
         gc_set->operator[](tile_group_id)[tuple_slot] = GCVersionType::COMMIT_DELETE;
-        log_manager.StartPersistTxn(100);
-        log_manager.LogDelete(ItemPointer(tile_group_id, tuple_slot));
-        log_manager.EndPersistTxn(100);
+        //log_manager.StartPersistTxn(end_commit_id);
+        log_manager.LogDelete(ItemPointer(tile_group_id, tuple_slot),5);
+        //log_manager.EndPersistTxn(end_commit_id);
       } else if (tuple_entry.second == RWType::INSERT) {
 
 
@@ -882,9 +879,9 @@ ResultType TimestampOrderingTransactionManager::CommitTransaction(
         tile_group_header->SetTransactionId(tuple_slot, INITIAL_TXN_ID);
 
         // nothing to be added to gc set.
-        log_manager.StartPersistTxn(100);
-        log_manager.LogInsert(ItemPointer(tile_group_id, tuple_slot));
-        log_manager.EndPersistTxn(100);
+        //log_manager.StartPersistTxn(100);
+        log_manager.LogInsert(ItemPointer(tile_group_id, tuple_slot),5);
+        //log_manager.EndPersistTxn(100);
 
 
       } else if (tuple_entry.second == RWType::INS_DEL) {
@@ -910,10 +907,6 @@ ResultType TimestampOrderingTransactionManager::CommitTransaction(
   }
 
   ResultType result = current_txn->GetResult();
-
-  //log_manager.DeregisterWorker();
-  //log_manager.DeregisterWorker();
-  //log_manager.End
 
   EndTransaction(current_txn);
