+++ src/include/logging/log_manager_factory.h
@@ -13,9 +13,8 @@
 #pragma once
 
 #include "logging/log_manager.h"
-#include "logging/reordered_phylog_log_manager.h"
+#include "logging/wal_log_manager.h"
 #include "logging/dummy_log_manager.h"
-//#include "logging/logical_log_manager.h"
 
 namespace peloton {
 namespace logging {
@@ -27,7 +26,7 @@ class LogManagerFactory {
     switch (logging_type_) {
 
       case LoggingType::ON:
-        return ReorderedPhyLogLogManager::GetInstance();
+        return WalLogManager::GetInstance();
       
       default:
         return DummyLogManager::GetInstance();
