+++ src/logging/wal_log_manager.cpp
@@ -12,8 +12,7 @@
 
 #include <cstdio>
 
-#include "concurrency/epoch_manager_factory.h"
-#include "logging/reordered_phylog_log_manager.h"
+#include "logging/wal_log_manager.h"
 #include "logging/durability_factory.h"
 #include "catalog/manager.h"
 #include "common/container_tuple.h"
@@ -23,7 +22,7 @@
 namespace peloton {
 namespace logging {
 
-void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
+void WalLogManager::WriteRecordToBuffer(LogRecord &record) {
   if(likely_branch(is_running_)){
 
   // Reset the output buffer
@@ -44,6 +43,7 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
       auto tg = manager.GetTileGroup(tuple_pos.block).get();
       std::vector<catalog::Column> columns;
 
+      output_buffer_.WriteLong(record.GetCommitId());
       // Write down the database id and the table id
       output_buffer_.WriteLong(tg->GetDatabaseId());
       output_buffer_.WriteLong(tg->GetTableId());
@@ -74,27 +74,12 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
       auto &manager = catalog::Manager::GetInstance();
       auto tuple_pos = record.GetItemPointer();
       auto tg = manager.GetTileGroup(tuple_pos.block).get();
-      //std::vector<catalog::Column> columns;
 
+      output_buffer_.WriteLong(record.GetCommitId());
       // Write down the database id and the table id
       output_buffer_.WriteLong(tg->GetDatabaseId());
       output_buffer_.WriteLong(tg->GetTableId());
 
-
-      // Write the full tuple into the buffer
-      /*for(auto schema : tg->GetTileSchemas()){
-          for(auto column : schema.GetColumns()){
-              columns.push_back(column);
-          }
-      }
-
-      ContainerTuple<storage::TileGroup> container_tuple(
-        tg, tuple_pos.offset
-      );
-      for(oid_t oid = 0; oid < columns.size(); oid++){
-        auto val = container_tuple.GetValue(oid);
-        val.SerializeTo(output_buffer_);
-      }*/
       output_buffer_.WriteLong(tuple_pos.block);
       output_buffer_.WriteLong(tuple_pos.offset);
 
@@ -109,6 +94,7 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
       auto tg = manager.GetTileGroup(tuple_pos.block).get();
       std::vector<catalog::Column> columns;
 
+      output_buffer_.WriteLong(record.GetCommitId());
       // Write down the database id and the table id
       output_buffer_.WriteLong(tg->GetDatabaseId());
       output_buffer_.WriteLong(tg->GetTableId());
@@ -116,7 +102,8 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
       output_buffer_.WriteLong(old_tuple_pos.block);
       output_buffer_.WriteLong(old_tuple_pos.offset);
 
-
+      output_buffer_.WriteLong(tuple_pos.block);
+      output_buffer_.WriteLong(tuple_pos.offset);
       // Write the full tuple into the buffer
       for(auto schema : tg->GetTileSchemas()){
           for(auto column : schema.GetColumns()){
@@ -158,10 +145,9 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
   bool is_success = buffer_ptr_->WriteData(output_buffer_.Data(), output_buffer_.Size());
   if (is_success == false) {
     // A buffer is full, pass it to the front end logger
-    // Get a new buffer and register it to current epoch
+    // Get a new buffer
     logger_->PersistLogBuffer(buffer_ptr_);
     buffer_ptr_ = new LogBuffer();
-    //buffer_ptr = RegisterNewBufferToEpoch(std::move(GetBuffer()));
     // Write it again
     is_success = buffer_ptr_->WriteData(output_buffer_.Data(), output_buffer_.Size());
     PL_ASSERT(is_success);
@@ -170,49 +156,49 @@ void ReorderedPhyLogLogManager::WriteRecordToBuffer(LogRecord &record) {
 }
 
 
-void ReorderedPhyLogLogManager::StartPersistTxn(cid_t current_cid) {
+void WalLogManager::StartPersistTxn(cid_t current_cid) {
   // Log down the begin of a transaction
   LogRecord record = LogRecordFactory::CreateTxnRecord(LogRecordType::TRANSACTION_BEGIN, current_cid);
   WriteRecordToBuffer(record);
 }
 
-void ReorderedPhyLogLogManager::EndPersistTxn(cid_t current_cid) {
-//  PL_ASSERT(tl_worker_ctx);
+void WalLogManager::EndPersistTxn(cid_t current_cid) {
   LogRecord record = LogRecordFactory::CreateTxnRecord(LogRecordType::TRANSACTION_COMMIT, current_cid);
   WriteRecordToBuffer(record);
-//  tl_worker_ctx->current_commit_eid = MAX_EID;
 }
 
 
-void ReorderedPhyLogLogManager::LogInsert(const ItemPointer &tuple_pos) {
-  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_INSERT, tuple_pos);
+void WalLogManager::LogInsert(const ItemPointer &tuple_pos, cid_t current_cid) {
+  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_INSERT, tuple_pos, current_cid);
   WriteRecordToBuffer(record);
+  logger_->PersistLogBuffer(buffer_ptr_);
 }
 
-void ReorderedPhyLogLogManager::LogUpdate(const ItemPointer &tuple_old_pos, const ItemPointer &tuple_pos) {
-  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_UPDATE, tuple_pos);
+void WalLogManager::LogUpdate(const ItemPointer &tuple_old_pos, const ItemPointer &tuple_pos, cid_t current_cid) {
+  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_UPDATE, tuple_pos, current_cid);
   record.SetOldItemPointer(tuple_old_pos);
   WriteRecordToBuffer(record);
+  logger_->PersistLogBuffer(buffer_ptr_);
 }
 
-void ReorderedPhyLogLogManager::LogDelete(const ItemPointer &tuple_pos_deleted) {
-  // Need the tuple value for the deleted tuple
-  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_DELETE, tuple_pos_deleted);
+void WalLogManager::LogDelete(const ItemPointer &tuple_pos_deleted, cid_t current_cid) {
+  LogRecord record = LogRecordFactory::CreateTupleRecord(LogRecordType::TUPLE_DELETE, tuple_pos_deleted,current_cid);
   WriteRecordToBuffer(record);
+  logger_->PersistLogBuffer(buffer_ptr_);
 }
 
-void ReorderedPhyLogLogManager::DoRecovery(){
+void WalLogManager::DoRecovery(){
     logger_->StartRecovery();
     logger_->WaitForRecovery();
  }
 
-void ReorderedPhyLogLogManager::StartLoggers() {
+void WalLogManager::StartLoggers() {
   logger_->SetLogBuffer(buffer_ptr_);
   logger_->StartLogging();
   is_running_ = true;
 }
 
-void ReorderedPhyLogLogManager::StopLoggers() {
+void WalLogManager::StopLoggers() {
   logger_->StopLogging();
   is_running_ = false;
