+++ test/type/serialize_test.cpp
@@ -194,10 +194,10 @@ TEST_F(SerializeTests, SerializeTupleToFileTest) {
     auto tuple2 = storage::Tuple(schema, true);
     tuple2.DeserializeWithHeaderFrom(record_decode);
 
+    EXPECT_EQ(0,tuple.Compare(tuple2));
     delete schema;
     delete buf;
     delete new_file_handle;
-    EXPECT_EQ(0,tuple.Compare(tuple2));
 }
 
 }  // End test namespace
